Fix parameters for interactive governor
Two commits to the kernel broke the interactive governor. They resulted in having both cores being stuck online and at a minimum frequency of 1026 MHz at all times while the screen was turned on.

The interactive governor was updated in two commits;
https://github.com/TeamApexQ/android_kernel_samsung_d2/commit/e12a8e83d922eb3800dd459dd3c3068bde8a04c2
https://github.com/TeamApexQ/android_kernel_samsung_d2/commit/c42353f0e809472b0cb2dd5831d61340debbe559

Effectively, what they do, is test if any online core's load exceeds up_threshold_any_cpu_load, and if any (could be other) core's frequency exceeds up_threshold_any_cpu_freq. If both of these conditions are true, it forces ALL CPU cores up to sync_freq, which in our case was set to 960 MHz, which does not exist in our frequency tables. It would pick the next frequency up, which was 1026 MHz. We also update the sync_freq to reflect this.